# backend/Contratos/openapi/ver catalogo/openapi.json
{"openapi":"3.1.0","info":{"title":"Chapinflix - Ver Catálogo","description":"Microservicio para visualización del catálogo de películas","version":"0.1.0"},"paths":{"/catalog/most-popular":{"get":{"tags":["catalog-viewer"],"summary":"Get Most Popular","description":"Obtiene las películas más populares","operationId":"get_most_popular_catalog_most_popular_get","parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"default":20,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CarouselResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/catalog/top-15":{"get":{"tags":["catalog-viewer"],"summary":"Get Top 15","description":"Obtiene el Top 15 de películas","operationId":"get_top_15_catalog_top_15_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CarouselResponse"}}}}}}},"/catalog/recently-added":{"get":{"tags":["catalog-viewer"],"summary":"Get Recently Added","description":"Obtiene las películas recientemente agregadas","operationId":"get_recently_added_catalog_recently_added_get","parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"default":20,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CarouselResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/catalog/recently-watched":{"get":{"tags":["catalog-viewer"],"summary":"Get Recently Watched","description":"Obtiene las películas vistas recientemente por el usuario","operationId":"get_recently_watched_catalog_recently_watched_get","parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"default":20,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CarouselResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/catalog/watch-again":{"get":{"tags":["catalog-viewer"],"summary":"Get Watch Again","description":"Obtiene las películas para ver otra vez (vistas >1 vez)","operationId":"get_watch_again_catalog_watch_again_get","parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"default":20,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CarouselResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/catalog/categories":{"get":{"tags":["catalog-viewer"],"summary":"Get Categories","description":"Obtiene todas las categorías disponibles","operationId":"get_categories_catalog_categories_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/Category"},"type":"array","title":"Response Get Categories Catalog Categories Get"}}}}}}},"/catalog/category/{slug}":{"get":{"tags":["catalog-viewer"],"summary":"Get Movies By Category","description":"Obtiene películas de una categoría específica","operationId":"get_movies_by_category_catalog_category__slug__get","parameters":[{"name":"slug","in":"path","required":true,"schema":{"type":"string","title":"Slug"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"default":20,"title":"Limit"}},{"name":"offset","in":"query","required":false,"schema":{"type":"integer","minimum":0,"default":0,"title":"Offset"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CarouselResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/catalog/featured":{"get":{"tags":["catalog-viewer"],"summary":"Get Featured Movie","description":"Obtiene la película destacada para el encabezado","operationId":"get_featured_movie_catalog_featured_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"anyOf":[{"$ref":"#/components/schemas/FeaturedMovie"},{"type":"null"}],"title":"Response Get Featured Movie Catalog Featured Get"}}}}}}},"/health":{"get":{"summary":"Health Check","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"CarouselResponse":{"properties":{"items":{"items":{"$ref":"#/components/schemas/MovieCarouselItem"},"type":"array","title":"Items"},"total":{"type":"integer","title":"Total"}},"type":"object","required":["items","total"],"title":"CarouselResponse"},"Category":{"properties":{"category_id":{"type":"string","title":"Category Id"},"name":{"type":"string","title":"Name"},"slug":{"type":"string","title":"Slug"},"parent_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Parent Id"}},"type":"object","required":["category_id","name","slug","parent_id"],"title":"Category"},"FeaturedMovie":{"properties":{"movie_id":{"type":"string","title":"Movie Id"},"title":{"type":"string","title":"Title"},"slug":{"type":"string","title":"Slug"},"synopsis_short":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Synopsis Short"},"synopsis_long":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Synopsis Long"},"banner_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Banner Url"},"poster_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Poster Url"},"classification_code":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Classification Code"},"duration_minutes":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Duration Minutes"},"is_free":{"type":"boolean","title":"Is Free"}},"type":"object","required":["movie_id","title","slug","synopsis_short","synopsis_long","banner_url","poster_url","classification_code","duration_minutes","is_free"],"title":"FeaturedMovie"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"MovieCarouselItem":{"properties":{"movie_id":{"type":"string","title":"Movie Id"},"title":{"type":"string","title":"Title"},"slug":{"type":"string","title":"Slug"},"synopsis_short":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Synopsis Short"},"poster_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Poster Url"},"classification_code":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Classification Code"},"duration_minutes":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Duration Minutes"},"is_free":{"type":"boolean","title":"Is Free"},"view_count":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"View Count"},"upload_date":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Upload Date"},"last_viewed":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Last Viewed"},"watch_count":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Watch Count"}},"type":"object","required":["movie_id","title","slug","synopsis_short","poster_url","classification_code","duration_minutes","is_free"],"title":"MovieCarouselItem"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}